moved LibtonodeEnvironment to testutils
